+++ app/Http/Controllers/RoleController.php
@@ -7,6 +7,9 @@ use Illuminate\Http\Request;
 use App\Http\Requests\StoreRoleRequest;
 use App\Repository\RepositoryInterface;
 use App\Http\Requests\UpdateRoleRequest;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -35,6 +38,9 @@ class RoleController extends Controller
      */
     public function store(StoreRoleRequest $request)
     {
+        if ($request->user()->cannot('create', Role::class)) {
+            abort(403);
+        }
         return $this->repository->add($request);
     }
     
@@ -56,6 +62,9 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, string $id)
     {
+        if ($request->user()->cannot('update', Role::class)) {
+            abort(403);
+        }
         return $this->repository->update($request,$id);
     }
 
@@ -69,6 +78,9 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        if(Auth::user()->cannot('delete', Type::class)){
+            abort(403); 
+        }
         return $this->repository->delete($id);
     }
 }
