+++ app/Http/Controllers/TypeController.php
@@ -6,6 +6,10 @@ use Illuminate\Http\Request;
 use App\Http\Requests\StoreTypeRequest;
 use App\Repository\RepositoryInterface;
 use App\Http\Requests\UpdateTypeRequest;
+use App\Models\Type;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class TypeController extends Controller
 {    
@@ -20,7 +24,6 @@ class TypeController extends Controller
        
     }    
       
-      
     /**
      * index
      *
@@ -31,7 +34,6 @@ class TypeController extends Controller
         return $this->repository->searchBy($request);
     }
 
-  
     /**
      * store
      *
@@ -40,6 +42,9 @@ class TypeController extends Controller
      */
     public function store(StoreTypeRequest $request)
     {
+        if ($request->user()->cannot('update', Type::class)) {
+            abort(403);
+        }
         return $this->repository->add($request);
     }
 
@@ -51,7 +56,6 @@ class TypeController extends Controller
     public function show(string $id){
         return $this->repository->show($id);
     }
-   
 
     /**
      * update
@@ -62,9 +66,11 @@ class TypeController extends Controller
      */
     public function update(UpdateTypeRequest $request, string $id)
     {
+        if ($request->user()->cannot('update', Type::class)) {
+            abort(403);
+        }
         return $this->repository->update($request,$id);
     }
-
         
     /**
      * destroy
@@ -75,7 +81,9 @@ class TypeController extends Controller
      */
     public function destroy(string $id)
     {
+        if(Auth::user()->cannot('delete', Type::class)){
+            abort(403); 
+        }
         return $this->repository->delete($id);
     }
-
 }
