+++ app/Http/Controllers/WpUserController.php
@@ -7,8 +7,11 @@ use App\Http\Requests\StoreWpUserRequest;
 use App\Http\Requests\UpdateWpUserRequest;
 use App\Models\WpUser;
 use App\Services\WpUserService;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class WpUserController extends Controller
 {
@@ -40,6 +43,9 @@ class WpUserController extends Controller
      */
     public function store(StoreWpUserRequest $request): JsonResponse
     {
+        if ($request->user()->cannot('createWpUser', WpUser::class)) {
+            abort(403);
+        }
         return  $this->wpUserService->storeWpUser($request->all());
     }
         
@@ -62,6 +68,9 @@ class WpUserController extends Controller
      */
     public function update(UpdateWpUserRequest $request, WpUser $wpUser): JsonResponse
     { 
+        if ($request->user()->cannot('updateWpUser', WpUser::class)) {
+            abort(403);
+        }
         return  $this->wpUserService->updateWpUser($request->all(), $wpUser);
     }
     
@@ -73,6 +82,9 @@ class WpUserController extends Controller
      */
     public function destroy(WpUser $wpUser): JsonResponse
     {
+        if(Auth::user()->cannot('delete', WpUser::class)){
+            abort(403); 
+        }
         return$this->wpUserService->deleteWpUser($wpUser);
     }
 }
